MyBatis, Feign and RabbitMQ added

+++ order-service/src/main/java/ru/liga/orderservice/service/OrderItemService.java
@@ -2,39 +2,56 @@ package ru.liga.orderservice.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.annotation.Validated;
 import ru.liga.orderservice.dto.ItemToAddDTO;
 import ru.liga.orderservice.dto.OrderToCreateDTO;
 import ru.liga.orderservice.entity.Order;
 import ru.liga.orderservice.entity.OrderItem;
+import ru.liga.orderservice.entity.RestaurantMenuItem;
+import ru.liga.orderservice.exceptions.NoSuchOrderException;
 import ru.liga.orderservice.repository.OrderItemRepository;
 import ru.liga.orderservice.repository.RestaurantMenuItemRepository;
 
+import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
+@Validated
 public class OrderItemService {
     private final OrderItemRepository repository;
     private final RestaurantMenuItemRepository restaurantMenuItemRepository;
-    public List<OrderItem> addNewOrderItems(OrderToCreateDTO dto, Order order) {
+    public List<OrderItem> addNewOrderItems(@Valid OrderToCreateDTO dto, Order order) {
         List<OrderItem> orderItems = new ArrayList<>();
-        for(ItemToAddDTO i : dto.getMenuItems()) {
+        for(ItemToAddDTO itemToAddDTO : dto.getMenuItems()) {
             OrderItem item = new OrderItem();
-            item.setQuantity(i.getQuantity());
-            item.setRestaurantMenuItem(restaurantMenuItemRepository.getRestaurantMenuItemById(i.getMenuItemId()));
-            item.setPrice(restaurantMenuItemRepository.getRestaurantMenuItemById(i.getMenuItemId()).getPrice()
-                    * i.getQuantity());
-            item.setOrder(order);
-            orderItems.add(item);
-            repository.save(item);
+            item.setQuantity(itemToAddDTO.getQuantity());
+            if(restaurantMenuItemRepository.getRestaurantMenuItemById(itemToAddDTO.getMenuItemId()).isPresent()) {
+                RestaurantMenuItem resItem = restaurantMenuItemRepository.getRestaurantMenuItemById(itemToAddDTO.getMenuItemId()).get();
+                item.setRestaurantMenuItem(resItem);
+                item.setPrice(resItem.getPrice()
+                        * itemToAddDTO.getQuantity());
+                item.setOrder(order);
+                orderItems.add(item);
+                repository.save(item);
+            }
         }
         return orderItems;
     }
     public OrderItem getItemById(long id){
-        return repository.getOrderItemById(id);
+        if(repository.getOrderItemById(id).isPresent()) {
+            return repository.getOrderItemById(id).get();
+        } else {
+            return null;
+        }
     }
     public void deleteItemById(long id) {
-        repository.deleteById(id);
+        if(repository.getOrderItemById(id).isPresent()) {
+            repository.deleteById(id);
+        } else {
+            throw new NoSuchOrderException("There is no Order Item with id " + id);
+        }
+
     }
 }
